+++ torch/_inductor/fx_passes/joint_graph.py
@@ -8,12 +8,10 @@ from typing import Any, Dict, List, Set, Union
 import torch
 import torch._guards
 import torch.utils._pytree as pytree
-from torch._dynamo.utils import detect_fake_mode
 from torch._inductor.constant_folding import ConstantFolder
 from torch._inductor.fx_passes.dedupe_symint_uses import _SymHashingDict
 from torch.fx.experimental.symbolic_shapes import statically_known_true
 from torch.fx.passes.graph_transform_observer import GraphTransformObserver
-from torch.fx.passes.tensorify_python_scalars import tensorify_python_scalars
 from torch.multiprocessing.reductions import StorageWeakRef
 
 from ...utils._ordered_set import OrderedSet
@@ -460,10 +458,6 @@ def joint_graph_passes(graph: torch.fx.GraphModule):
     if config.pattern_matcher:
         for patterns in pass_patterns:
             count += patterns.apply(graph.graph)  # type: ignore[arg-type]
-
-    fake_mode = detect_fake_mode()
-    if fake_mode is not None:
-        tensorify_python_scalars(graph, fake_mode.shape_env)
 
     if not config.fallback_random:
         count += replace_random_passes(graph)
